+++ OneBlink.SDK/models/ApprovalsFormsApp.cs
@@ -1,0 +1,7 @@
+namespace OneBlink.SDK.Model
+{
+    public class ApprovalsFormsApp : FormsAppBase
+    {
+        public FormsAppStylesBase styles {get;set;}
+    }
+}