+++ OneBlink.SDK/models/FormsListFormsApp.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormsListFormsApp : FormsAppBase
+    {
+        public string slug {get;set;}
+        public List<long> formIds {get;set;}
+        public FormsListStyles styles {get;set;}
+    }
+}