+++ OneBlink.SDK/models/TilesFormsApp.cs
@@ -1,0 +1,8 @@
+namespace OneBlink.SDK.Model
+{
+    public class TilesFormsApp : FormsAppBase
+    {
+        public string slug {get;set;}
+        public FormsListStyles styles {get;set;}
+    }
+}