+++ OneBlink.SDK/models/VolunteerFormsAppCategories.cs
@@ -1,0 +1,7 @@
+namespace OneBlink.SDK.Model
+{
+    public class VolunteersFormsAppCategories
+    {
+        public string label {get;set;}
+    }
+}